chore: zero out -Wextra/-Wpedantic/-Wshadow/-Wconversion warnings
Bring the build to zero warnings under the project's strict flag set
(-Wall -Wextra -Wpedantic -Wshadow -Wconversion):

- jk_proto.c: explicit (float) casts on scaled uint32/int32 sensor values;
  (uint32_t) casts on the size_t drop-counter addends.
- classic_codec.c: (uint8_t)/(float) casts on decoded byte/scaled fields.
- dashboard.c, pack_view.c: wrap vk_widget_hide()'s (state & ~VK_STATE_*)
  in the intended (uint32_t) cast (the libviper macro narrows u64->u32).
- pack_view.c: drop unused mk_lab_txt/mk_meter; reuse the outer clientH
  instead of shadowing it.
- config.{c,h}: guard the redundant _DEFAULT_SOURCE #define (already set via
  -D_GNU_SOURCE) with #ifndef.
- mf_backtrace.c, test_config.c: consume write()/fread() results ((void)
  does not silence warn_unused_result).
- -Wformat-truncation: return-check where truncation is an error (config
  path resolve, host:port parse, device-settings JSON piece); bound via
  precision / buffer sizing where truncation is benign (layout caption,
  usb dev path, hostport, test path).

Behavior-neutral apart from: an over-long config value / host / path is now
rejected or skipped rather than silently truncated. No change to decode
logic. Pre-existing test failures (xd_plugin, usb_id, jk_plugin) unchanged.

+++ src/classic/classic_codec.c
@@ -28,7 +28,7 @@ classic_result_t classic_decode(const classic_registers_t *input,
     /* Ibatt: wire 4116, i16 /10 A (signed) */
     R(CLASSIC_REG_IBATT);
     out->ibatt_raw = I16_SIGNED(CLASSIC_REG_IBATT);
-    out->ibatt_a = out->ibatt_raw * 0.1f;
+    out->ibatt_a = (float)out->ibatt_raw * 0.1f;
 
     /* kWh today: wire 4117, u16 /10 kWh */
     R(CLASSIC_REG_KWH_TODAY);
@@ -40,7 +40,7 @@ classic_result_t classic_decode(const classic_registers_t *input,
 
     /* Charge stage: wire 4119, MSB = stage code */
     R(CLASSIC_REG_CHARGE_STAGE);
-    out->charge_stage_msb = (U16(CLASSIC_REG_CHARGE_STAGE) >> 8) & 0xFF;
+    out->charge_stage_msb = (uint8_t)((U16(CLASSIC_REG_CHARGE_STAGE) >> 8) & 0xFF);
     out->charge_stage_name[0] = '\0';
     const char *stage_str = classic_charge_stage_name(out->charge_stage_msb);
     strncpy(out->charge_stage_name, stage_str, sizeof(out->charge_stage_name) - 1);
@@ -71,12 +71,12 @@ classic_result_t classic_decode(const classic_registers_t *input,
     R(CLASSIC_REG_UNIT_MAC_LO);
     R(CLASSIC_REG_UNIT_MAC_MID);
     R(CLASSIC_REG_UNIT_MAC_HI);
-    out->unit_mac[0] = (U16(CLASSIC_REG_UNIT_MAC_HI)     >> 8) & 0xFF;
-    out->unit_mac[1] =  U16(CLASSIC_REG_UNIT_MAC_HI)      & 0xFF;
-    out->unit_mac[2] = (U16(CLASSIC_REG_UNIT_MAC_MID)    >> 8) & 0xFF;
-    out->unit_mac[3] =  U16(CLASSIC_REG_UNIT_MAC_MID)      & 0xFF;
-    out->unit_mac[4] = (U16(CLASSIC_REG_UNIT_MAC_LO)     >> 8) & 0xFF;
-    out->unit_mac[5] =  U16(CLASSIC_REG_UNIT_MAC_LO)       & 0xFF;
+    out->unit_mac[0] = (uint8_t)((U16(CLASSIC_REG_UNIT_MAC_HI)     >> 8) & 0xFF);
+    out->unit_mac[1] = (uint8_t)( U16(CLASSIC_REG_UNIT_MAC_HI)      & 0xFF);
+    out->unit_mac[2] = (uint8_t)((U16(CLASSIC_REG_UNIT_MAC_MID)    >> 8) & 0xFF);
+    out->unit_mac[3] = (uint8_t)( U16(CLASSIC_REG_UNIT_MAC_MID)      & 0xFF);
+    out->unit_mac[4] = (uint8_t)((U16(CLASSIC_REG_UNIT_MAC_LO)     >> 8) & 0xFF);
+    out->unit_mac[5] = (uint8_t)( U16(CLASSIC_REG_UNIT_MAC_LO)       & 0xFF);
 
     /* UNIT_Device_ID: wire[4111]<<16 | wire[4110] */
     R(CLASSIC_REG_DEVICE_ID_LO);
